pyproject.toml
[project]
name = "promptune"
version = "0.1.0"
description = "Promptune - Beam Search Prompt Tuning"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Afzal Mukhtar"}
]
keywords = ["prompt-optimization", "llm", "beam-search", "mcp"]

dependencies = [
    "pydantic>=2.0",
    "fastmcp>=0.4.0",
    "litellm>=1.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
azure = [
    "openai>=1.0",
]
ollama = [
    "ollama>=0.1",
]
all = [
    "openai>=1.0",
    "ollama>=0.1",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["schemas", "mcp_servers"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
